Revert "move ExampleTestUtil to examples/test and add error messages"
This reverts commit 0c72cb8531ce05123544c8c64ffd720084357e88.
