+++ examples/beam/src/test/java/edu/snu/nemo/examples/beam/BroadcastITCase.java
@@ -17,6 +17,7 @@ package edu.snu.nemo.examples.beam;
 
 import edu.snu.nemo.client.JobLauncher;
 import edu.snu.nemo.common.test.ArgBuilder;
+import edu.snu.nemo.common.test.ExampleTestUtil;
 import edu.snu.nemo.examples.beam.policy.DefaultPolicyParallelismFive;
 import edu.snu.nemo.examples.beam.policy.TransientResourcePolicyParallelismFive;
 import org.junit.After;
@@ -54,9 +55,9 @@ public final class BroadcastITCase {
   @After
   public void tearDown() throws Exception {
     try {
-      BeamExampleTestUtil.ensureOutputValidity(fileBasePath, outputFileName, expectedOutputFileName);
+      ExampleTestUtil.ensureOutputValidity(fileBasePath, outputFileName, expectedOutputFileName);
     } finally {
-      BeamExampleTestUtil.deleteOutputFile(fileBasePath, outputFileName);
+      ExampleTestUtil.deleteOutputFile(fileBasePath, outputFileName);
     }
   }
